port tests to .ts.
