+++ .gitignore
@@ -12,3 +12,5 @@ index.js
 index.js.map
 lib/**/*.js
 lib/**/*.js.map
+test/**/*.js
+test/**/*.js.map
