+++ .nycrc
@@ -1,0 +1,3 @@
+{
+    "extends": "@istanbuljs/nyc-config-typescript"
+}
