+++ package.json
@@ -7,8 +7,8 @@
     "Chris Troutner <chris.troutner@gmail.com>"
   ],
   "scripts": {
-    "test": "TEST=unit nyc --reporter=text mocha --require babel-core/register --timeout 30000 test/unit",
-    "test:integration": "TEST=integration nyc --reporter=text mocha --require babel-core/register --timeout 10000 test/integration",
+    "test": "nyc mocha test/unit --timeout 30000",
+    "test:integration": "TEST=integration nyc --reporter=text mocha --require babel-core/register --timeout 30000 test/integration",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "build": "tsc"
   },
@@ -34,9 +34,11 @@
     "@types/bip39": "^2.4.2",
     "@types/figlet": "^1.2.0",
     "@types/mkdirp": "^0.5.2",
+    "@types/mocha": "^5.2.6",
     "@types/node": "^10.11.7",
     "@types/node-emoji": "^1.8.1",
     "@types/randombytes": "^2.0.0",
+    "@types/sinon": "^7.0.11",
     "@types/touch": "^3.1.1",
     "@types/wif": "^2.0.1",
     "assert": "^1.4.1",
@@ -84,6 +86,7 @@
     "wif": "^2.0.6"
   },
   "devDependencies": {
+    "@istanbuljs/nyc-config-typescript": "^0.1.3",
     "chai": "^4.1.2",
     "coveralls": "^3.0.2",
     "eslint": "^5.5.0",
@@ -93,10 +96,12 @@
     "mocha": "^5.0.1",
     "nock": "^10.0.6",
     "node-mocks-http": "^1.7.0",
-    "nyc": "^13.3.0",
+    "nyc": "^14.1.1",
     "prettier": "^1.14.2",
     "semantic-release": "^15.13.3",
     "sinon": "^4.5.0",
+    "source-map-support": "^0.5.12",
+    "ts-node": "^8.1.0",
     "typescript": "^3.4.5"
   }
 }
