+++ test/mocha.opts
@@ -1,0 +1,4 @@
+--require ts-node/register #replace with ts-node/register/transpile-only if you have custom types
+--require source-map-support/register
+--recursivee
+test/unit/**/*.ts
