+++ test/unit/Control.ts
@@ -1,9 +1,11 @@
-const assert = require("assert")
-const axios = require("axios")
+import * as assert from "assert";
+import axios from "axios";
+import * as sinon from "sinon";
+
+// TODO: port from require to import syntax
 const BITBOX = require("../../lib/BITBOX").BITBOX
 const bitbox = new BITBOX()
 const Control = require("../../lib/Control").Control
-const sinon = require("sinon")
 const resturl = require("../../lib/BITBOX").resturl
 
 describe("#Control", () => {
@@ -20,7 +22,7 @@ describe("#Control", () => {
   })
 
   describe("#getInfo", () => {
-    let sandbox
+    let sandbox: any
     beforeEach(() => (sandbox = sinon.sandbox.create()))
     afterEach(() => sandbox.restore())
 
@@ -42,7 +44,7 @@ describe("#Control", () => {
       sandbox.stub(axios, "get").returns(resolved)
 
       bitbox.Control.getInfo()
-        .then(result => {
+        .then((result: any) => {
           assert.deepEqual(data, result)
         })
         .then(done, done)
@@ -50,7 +52,7 @@ describe("#Control", () => {
   })
 
   describe("#getMemoryInfo", () => {
-    let sandbox
+    let sandbox: any
     beforeEach(() => (sandbox = sinon.sandbox.create()))
     afterEach(() => sandbox.restore())
 
@@ -69,7 +71,7 @@ describe("#Control", () => {
       sandbox.stub(axios, "get").returns(resolved)
 
       bitbox.Control.getMemoryInfo()
-        .then(result => {
+        .then((result: any) => {
           assert.deepEqual(data, result)
         })
         .then(done, done)
