+++ test/unit/Generating.ts
@@ -1,9 +1,11 @@
-const assert = require("assert")
-const axios = require("axios")
+import * as assert from "assert";
+import axios from "axios";
+import * as sinon from "sinon";
+
+// TODO: port from require to import syntax
 const BITBOX = require("../../lib/BITBOX").BITBOX
 const bitbox = new BITBOX()
 const Generating = require("../../lib/Generating").Generating
-const sinon = require("sinon")
 const resturl = require("../../lib/BITBOX").resturl
 
 describe("#Generating", () => {
@@ -20,12 +22,12 @@ describe("#Generating", () => {
   })
 
   describe("#generateToAddress", () => {
-    let sandbox
+    let sandbox: any
     beforeEach(() => (sandbox = sinon.sandbox.create()))
     afterEach(() => sandbox.restore())
 
     it("should generate", done => {
-      const data = []
+      const data: any = []
       const resolved = new Promise(r => r({ data: data }))
       sandbox.stub(axios, "post").returns(resolved)
 
@@ -33,7 +35,7 @@ describe("#Generating", () => {
         1,
         "bitcoincash:qrdka2205f4hyukutc2g0s6lykperc8nsu5u2ddpqf"
       )
-        .then(result => {
+        .then((result: any) => {
           assert.deepEqual(data, result)
         })
         .then(done, done)
