+++ test/unit/RawTransactions.js
@@ -1,143 +1,179 @@
+"use strict";
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
+var _this = this;
+Object.defineProperty(exports, "__esModule", { value: true });
 /*
   TODO:
   -Replace old unit tests mocking axios with the more generalized nock library.
    See the sendRawTransaction test for an example.
   -Create a mocking library of data to compare unit and integration tests.
 */
-
-const assert = require("assert")
-const axios = require("axios")
-const BITBOX = require("../../lib/BITBOX").BITBOX
-const bitbox = new BITBOX()
-const RawTransactions = require("../../lib/RawTransactions").RawTransactions
-const resturl = require("../../lib/BITBOX").resturl
-
-const sinon = require("sinon")
-const nock = require("nock") // HTTP mocking
-
+var assert = require("assert");
+var axios_1 = require("axios");
+var sinon = require("sinon");
+// TODO: port from require to import syntax
+var BITBOX = require("../../lib/BITBOX").BITBOX;
+var bitbox = new BITBOX();
+var RawTransactions = require("../../lib/RawTransactions").RawTransactions;
+var resturl = require("../../lib/BITBOX").resturl;
+var nock = require("nock"); // HTTP mocking
 // Used for debugging
-const util = require("util")
-util.inspect.defaultOptions = { depth: 1 }
-
-describe("#RawTransactions", () => {
-  describe("#PriceConstructor", () => {
-    it("should create instance of RawTransactions", () => {
-      const rawtransactions = new RawTransactions()
-      assert.equal(rawtransactions instanceof RawTransactions, true)
-    })
-
-    it("should have a restURL property", () => {
-      const rawtransactions = new RawTransactions()
-      assert.equal(rawtransactions.restURL, resturl)
-    })
-  })
-
-  describe("#decodeRawTransaction", () => {
-    let sandbox
-    beforeEach(() => (sandbox = sinon.sandbox.create()))
-    afterEach(() => sandbox.restore())
-
-    it("should decode raw transaction", done => {
-      const data = {
-        txid:
-          "4ebd325a4b394cff8c57e8317ccf5a8d0e2bdf1b8526f8aad6c8e43d8240621a",
-        hash:
-          "4ebd325a4b394cff8c57e8317ccf5a8d0e2bdf1b8526f8aad6c8e43d8240621a",
-        size: 10,
-        version: 2,
-        locktime: 0,
-        vin: [],
-        vout: []
-      }
-
-      const resolved = new Promise(r => r({ data: data }))
-      sandbox.stub(axios, "get").returns(resolved)
-
-      bitbox.RawTransactions.decodeRawTransaction(["02000000000000000000"])
-        .then(result => {
-          assert.deepEqual(data, result[0])
-        })
-        .then(done, done)
-    })
-  })
-
-  describe("#decodeScript", () => {
-    let sandbox
-    beforeEach(() => (sandbox = sinon.sandbox.create()))
-    afterEach(() => sandbox.restore())
-
-    it("should decode script", async () => {
-      const data = {
-        asm: "OP_RETURN 5361746f736869204e616b616d6f746f",
-        type: "nulldata",
-        p2sh: "bitcoincash:prswx5965nfumux9qng5kj8hw603vcne7q08t8c6jp"
-      }
-
-      const resolved = new Promise(r => r({ data: data }))
-      sandbox.stub(axios, "get").returns(resolved)
-
-      const result = await bitbox.RawTransactions.decodeScript(
-        "6a105361746f736869204e616b616d6f746f"
-      )
-      //console.log(`result: ${util.inspect(result)}`)
-
-      assert.deepEqual(data, result)
-    })
-  })
-
-  describe("#getRawTransaction", () => {
-    let sandbox
-    beforeEach(() => (sandbox = sinon.sandbox.create()))
-    afterEach(() => sandbox.restore())
-
-    it("should get raw transaction", done => {
-      const data =
-        "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
-
-      const resolved = new Promise(r => r({ data: data }))
-      sandbox.stub(axios, "get").returns(resolved)
-
-      bitbox.RawTransactions.getRawTransaction([
-        "808d617eccaad4f1397fe07a06ec5ed15a0821cf22a3e0931c0c92aef9e572b6"
-      ])
-        .then(result => {
-          assert.deepEqual(data, result[0])
-        })
-        .then(done, done)
-    })
-  })
-
-  describe("#sendRawTransaction", () => {
-    let sandbox
-    beforeEach(() => (sandbox = sinon.sandbox.create()))
-    afterEach(() => sandbox.restore())
-
-    it("should send single raw transaction", async () => {
-      const data = "Error: transaction already in block chain"
-
-      const resolved = new Promise(r => r({ data: data }))
-      sandbox.stub(axios, "get").returns(resolved)
-
-      const result = await bitbox.RawTransactions.sendRawTransaction(
-        "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
-      )
-
-      assert.equal(data, result)
-    })
-
-    it("should send an array of raw transactions", async () => {
-      const data = "Error: transaction already in block chain"
-
-      // Mock the http call to rest.bitcoin.com
-      nock(`${bitbox.RawTransactions.restURL}`)
-        .post(uri => uri.includes(`/`))
-        .reply(200, { data: data })
-
-      const result = await bitbox.RawTransactions.sendRawTransaction([
-        "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
-      ])
-
-      assert.equal(data, result.data)
-    })
-  })
-})
+var util = require("util");
+util.inspect.defaultOptions = { depth: 1 };
+describe("#RawTransactions", function () {
+    describe("#PriceConstructor", function () {
+        it("should create instance of RawTransactions", function () {
+            var rawtransactions = new RawTransactions();
+            assert.equal(rawtransactions instanceof RawTransactions, true);
+        });
+        it("should have a restURL property", function () {
+            var rawtransactions = new RawTransactions();
+            assert.equal(rawtransactions.restURL, resturl);
+        });
+    });
+    describe("#decodeRawTransaction", function () {
+        var sandbox;
+        beforeEach(function () { return (sandbox = sinon.sandbox.create()); });
+        afterEach(function () { return sandbox.restore(); });
+        it("should decode raw transaction", function (done) {
+            var data = {
+                txid: "4ebd325a4b394cff8c57e8317ccf5a8d0e2bdf1b8526f8aad6c8e43d8240621a",
+                hash: "4ebd325a4b394cff8c57e8317ccf5a8d0e2bdf1b8526f8aad6c8e43d8240621a",
+                size: 10,
+                version: 2,
+                locktime: 0,
+                vin: [],
+                vout: []
+            };
+            var resolved = new Promise(function (r) { return r({ data: data }); });
+            sandbox.stub(axios_1.default, "get").returns(resolved);
+            bitbox.RawTransactions.decodeRawTransaction(["02000000000000000000"])
+                .then(function (result) {
+                assert.deepEqual(data, result[0]);
+            })
+                .then(done, done);
+        });
+    });
+    describe("#decodeScript", function () {
+        var sandbox;
+        beforeEach(function () { return (sandbox = sinon.sandbox.create()); });
+        afterEach(function () { return sandbox.restore(); });
+        it("should decode script", function () { return __awaiter(_this, void 0, void 0, function () {
+            var data, resolved, result;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        data = {
+                            asm: "OP_RETURN 5361746f736869204e616b616d6f746f",
+                            type: "nulldata",
+                            p2sh: "bitcoincash:prswx5965nfumux9qng5kj8hw603vcne7q08t8c6jp"
+                        };
+                        resolved = new Promise(function (r) { return r({ data: data }); });
+                        sandbox.stub(axios_1.default, "get").returns(resolved);
+                        return [4 /*yield*/, bitbox.RawTransactions.decodeScript("6a105361746f736869204e616b616d6f746f")
+                            //console.log(`result: ${util.inspect(result)}`)
+                        ];
+                    case 1:
+                        result = _a.sent();
+                        //console.log(`result: ${util.inspect(result)}`)
+                        assert.deepEqual(data, result);
+                        return [2 /*return*/];
+                }
+            });
+        }); });
+    });
+    // describe("#getRawTransaction", () => {
+    //   let sandbox: any
+    //   beforeEach(() => (sandbox = sinon.sandbox.create()))
+    //   afterEach(() => sandbox.restore())
+    //   it("should get raw transaction", done => {
+    //     const data =
+    //       "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
+    //     const resolved = new Promise(r => r({ data: data }))
+    //     sandbox.stub(axios, "get").returns(resolved)
+    //     bitbox.RawTransactions.getRawTransaction([
+    //       "808d617eccaad4f1397fe07a06ec5ed15a0821cf22a3e0931c0c92aef9e572b6"
+    //     ])
+    //       .then((result: any) => {
+    //         assert.deepEqual(data, result[0])
+    //       })
+    //       .then(done, done)
+    //   })
+    // })
+    describe("#sendRawTransaction", function () {
+        var sandbox;
+        beforeEach(function () { return (sandbox = sinon.sandbox.create()); });
+        afterEach(function () { return sandbox.restore(); });
+        it("should send single raw transaction", function () { return __awaiter(_this, void 0, void 0, function () {
+            var data, resolved, result;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        data = "Error: transaction already in block chain";
+                        resolved = new Promise(function (r) { return r({ data: data }); });
+                        sandbox.stub(axios_1.default, "get").returns(resolved);
+                        return [4 /*yield*/, bitbox.RawTransactions.sendRawTransaction("020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000")];
+                    case 1:
+                        result = _a.sent();
+                        assert.equal(data, result);
+                        return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should send an array of raw transactions", function () { return __awaiter(_this, void 0, void 0, function () {
+            var data, result;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        data = "Error: transaction already in block chain";
+                        // Mock the http call to rest.bitcoin.com
+                        nock("" + bitbox.RawTransactions.restURL)
+                            .post(function (uri) { return uri.includes("/"); })
+                            .reply(200, { data: data });
+                        return [4 /*yield*/, bitbox.RawTransactions.sendRawTransaction([
+                                "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
+                            ])];
+                    case 1:
+                        result = _a.sent();
+                        assert.equal(data, result.data);
+                        return [2 /*return*/];
+                }
+            });
+        }); });
+    });
+});
+//# sourceMappingURL=RawTransactions.js.map
