+++ tsconfig.json
@@ -5,8 +5,15 @@
     "target": "es5",
     "downlevelIteration": true,
     "moduleResolution": "node",
-    "lib": ["es2017"],
+    "lib": [
+      "es2017"
+    ],
     "sourceMap": true
   },
-  "include": ["./lib", "./test", "index.ts"]
+  "include": [
+    "./lib",
+    "./test",
+    "./test/unit/**/*.ts",
+    "index.ts"
+  ]
 }
